refactor(server): split get_content into cohesive modules
The 1,971-line tool file becomes a 15-line entry shim re-exporting the same
surface, with schema/types/query/render/handler/watcher-mode modules under
tools/get_content/. Code moved verbatim (diff-audited); two any[] params
tightened where element types were provable. External import paths unchanged.
